requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
blinker==1.6.2
cffi==1.15.1
click==8.1.7
exceptiongroup==1.1.3
Flask==2.3.2
Flask-WTF==1.1.1
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.1
password-validator==1.0
pluggy==1.2.0
pycparser==2.21
pytest==7.4.0
python-dotenv==1.0.0
tomli==2.0.1
Werkzeug==2.3.7
WTForms==3.0.1
beautifulsoup4==4.12.2
SQLAlchemy~=2.0.21